requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.13.0
django-versatileimagefield==2.2
Pillow==8.4.0
psycopg2-binary==2.9.3
python-decouple==3.6
python-magic==0.4.27
pytz==2021.3
sqlparse==0.4.2
